requirements/prod.txt
djangae==0.9.6
django>=1.8,<1.10
git+https://github.com/mozilla/django-csp.git#egg=djangocsp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
six
croniter==0.3.13
django-crispy-forms==1.6.1
